+++ lib/types.ts
@@ -508,28 +508,43 @@ export type StickerEventMessage = {
 
 export type Postback = {
   data: string;
+  params?: DateTimePostback | RichMenuSwitchPostback;
+};
+
+/**
+ * Object with the date and time selected by a user through a
+ * [datetime picker action](https://developers.line.biz/en/reference/messaging-api/#datetime-picker-action).
+ * Only returned for postback actions via a
+ * [datetime picker action](https://developers.line.biz/en/reference/messaging-api/#datetime-picker-action).
+ * The `full-date`, `time-hour`, and `time-minute` formats follow the
+ * [RFC3339 protocol](https://www.ietf.org/rfc/rfc3339.txt).
+ */
+type DateTimePostback = {
   /**
-   * Object with the date and time selected by a user through a
-   * [datetime picker action](https://developers.line.biz/en/reference/messaging-api/#datetime-picker-action).
-   * Only returned for postback actions via a
-   * [datetime picker action](https://developers.line.biz/en/reference/messaging-api/#datetime-picker-action).
-   * The `full-date`, `time-hour`, and `time-minute` formats follow the
-   * [RFC3339 protocol](https://www.ietf.org/rfc/rfc3339.txt).
+   * Date selected by user. Only included in the `date` mode.
    */
-  params?: {
-    /**
-     * Date selected by user. Only included in the `date` mode.
-     */
-    date?: string;
-    /**
-     * Time selected by the user. Only included in the `time` mode.
-     */
-    time?: string;
-    /**
-     * Date and time selected by the user. Only included in the `datetime` mode.
-     */
-    datetime?: string;
-  };
+  date?: string;
+  /**
+   * Time selected by the user. Only included in the `time` mode.
+   */
+  time?: string;
+  /**
+   * Date and time selected by the user. Only included in the `datetime` mode.
+   */
+  datetime?: string;
+};
+
+/**
+ * Object with rich menu alias ID selected by user via rich menu switch action.
+ * https://developers.line.biz/en/reference/messaging-api/#postback-params-object-for-richmenu-switch-action
+ */
+type RichMenuSwitchPostback = {
+  newRichMenuAliasId: string;
+  status:
+    | "SUCCESS"
+    | "RICHMENU_ALIAS_ID_NOTFOUND"
+    | "RICHMENU_NOTFOUND"
+    | "FAILED";
 };
 
 /**
@@ -577,13 +592,25 @@ export type TextMessage = MessageCommon & {
   /**
    * Message text. You can include the following emoji:
    *
+   * - LINE emojis. Use a $ character as a placeholder and specify the product ID and emoji ID of the LINE emoji you want to use in the emojis property.
    * - Unicode emoji
-   * - LINE original emoji
+   * - (Deprecated) LINE original unicode emojis
    *   ([Unicode codepoint table for LINE original emoji](https://developers.line.biz/media/messaging-api/emoji-list.pdf))
    *
-   * Max: 2000 characters
+   * Max: 5000 characters
    */
   text: string;
+
+  /**
+   * One or more LINE emoji.
+   *
+   * Max: 20 LINE emoji
+   */
+  emojis?: {
+    index: number;
+    productId: string;
+    emojiId: string;
+  }[];
 };
 
 /**
@@ -592,7 +619,7 @@ export type TextMessage = MessageCommon & {
 export type ImageMessage = MessageCommon & {
   type: "image";
   /**
-   * Image URL (Max: 1000 characters)
+   * Image URL (Max: 2000 characters)
    *
    * - **HTTPS**
    * - JPEG
@@ -601,7 +628,7 @@ export type ImageMessage = MessageCommon & {
    */
   originalContentUrl: string;
   /**
-   * Preview image URL (Max: 1000 characters)
+   * Preview image URL (Max: 2000 characters)
    *
    * - **HTTPS**
    * - JPEG
@@ -617,7 +644,7 @@ export type ImageMessage = MessageCommon & {
 export type VideoMessage = MessageCommon & {
   type: "video";
   /**
-   * URL of video file (Max: 1000 characters)
+   * URL of video file (Max: 2000 characters)
    *
    * - **HTTPS**
    * - mp4
@@ -628,7 +655,7 @@ export type VideoMessage = MessageCommon & {
    */
   originalContentUrl: string;
   /**
-   * URL of preview image (Max: 1000 characters)
+   * URL of preview image (Max: 2000 characters)
    *
    * - **HTTPS**
    * - JPEG
@@ -644,7 +671,7 @@ export type VideoMessage = MessageCommon & {
 export type AudioMessage = MessageCommon & {
   type: "audio";
   /**
-   * URL of audio file (Max: 1000 characters)
+   * URL of audio file (Max: 2000 characters)
    *
    * - **HTTPS**
    * - m4a
@@ -702,7 +729,7 @@ export type ImageMapMessage = MessageCommon & {
   type: "imagemap";
   /**
    * [Base URL](https://developers.line.biz/en/reference/messaging-api/#base-url) of image
-   * (Max: 1000 characters, **HTTPS**)
+   * (Max: 2000 characters, **HTTPS**)
    */
   baseUrl: string;
   /**
@@ -715,7 +742,7 @@ export type ImageMapMessage = MessageCommon & {
    */
   video?: {
     /**
-     * URL of video file (Max: 1000 characters)
+     * URL of video file (Max: 2000 characters)
      *
      * - **HTTPS**
      * - mp4
@@ -726,7 +753,7 @@ export type ImageMapMessage = MessageCommon & {
      */
     originalContentUrl: string;
     /**
-     * URL of preview image (Max: 1000 characters)
+     * URL of preview image (Max: 2000 characters)
      *
      * - **HTTPS**
      * - JPEG
@@ -1281,7 +1308,7 @@ export type FlexFiller = {
 export type FlexIcon = {
   type: "icon";
   /**
-   * Image URL
+   * Image URL (Max character limit: 2000)
    *
    * Protocol: HTTPS
    * Image format: JPEG or PNG
@@ -1334,7 +1361,7 @@ export type FlexIcon = {
 export type FlexImage = {
   type: "image";
   /**
-   * Image URL
+   * Image URL (Max character limit: 2000)
    *
    * - Protocol: HTTPS
    * - Image format: JPEG or PNG
@@ -1683,7 +1710,7 @@ export type TemplateContent =
 export type TemplateButtons = {
   type: "buttons";
   /**
-   * Image URL (Max: 1000 characters)
+   * Image URL (Max: 2000 characters)
    *
    * - HTTPS
    * - JPEG or PNG
@@ -1797,7 +1824,7 @@ export type TemplateCarousel = {
 
 export type TemplateColumn = {
   /**
-   * Image URL (Max: 1000 characters)
+   * Image URL (Max: 2000 characters)
    *
    * - HTTPS
    * - JPEG or PNG
@@ -1846,7 +1873,7 @@ export type TemplateImageCarousel = {
 
 export type TemplateImageColumn = {
   /**
-   * Image URL (Max: 1000 characters)
+   * Image URL (Max: 2000 characters)
    *
    * - HTTPS
    * - JPEG or PNG
@@ -1916,6 +1943,7 @@ export type QuickReplyItem = {
    * - [Camera action](https://developers.line.biz/en/reference/messaging-api/#camera-action)
    * - [Camera roll action](https://developers.line.biz/en/reference/messaging-api/#camera-roll-action)
    * - [Location action](https://developers.line.biz/en/reference/messaging-api/#location-action)
+   * - [URI action](https://developers.line.biz/en/reference/messaging-api/#uri-action)
    */
   action: Action;
 };
@@ -1944,6 +1972,7 @@ export type Sender = {
  * - [Message action](https://developers.line.biz/en/reference/messaging-api/#message-action)
  * - [URI action](https://developers.line.biz/en/reference/messaging-api/#uri-action)
  * - [Datetime picker action](https://developers.line.biz/en/reference/messaging-api/#datetime-picker-action)
+ * - [Rich menu switch action](https://developers.line.biz/en/reference/messaging-api/#richmenu-switch-action)
  * - [Camera action](https://developers.line.biz/en/reference/messaging-api/#camera-action)
  * - [Camera roll action](https://developers.line.biz/en/reference/messaging-api/#camera-roll-action)
  * - [Location action](https://developers.line.biz/en/reference/messaging-api/#location-action)
@@ -1953,6 +1982,7 @@ export type Action<ExtraFields = { label: string }> = (
   | MessageAction
   | URIAction
   | DatetimePickerAction
+  | RichMenuSwitchAction
   | { type: "camera" }
   | { type: "cameraRoll" }
   | { type: "location" }
@@ -2079,6 +2109,28 @@ export type DatetimePickerAction = {
 export type Size = {
   width: number;
   height: number;
+};
+
+/**
+ * When a control associated with this action is tapped, the URI specified in
+ * the `uri` property is opened.
+ */
+export type RichMenuSwitchAction = {
+  type: "richmenuswitch";
+  /**
+   * Action label. Optional for rich menus. Read when the user's device accessibility feature is enabled.
+   * Max character limit: 20. Supported on LINE for iOS 8.2.0 or later.
+   */
+  label?: string;
+  /**
+   * Rich menu alias ID to switch to.
+   */
+  richMenuAliasId: string;
+  /**
+   * String returned by the postback.data property of the postback event via a webhook
+   * Max character limit: 300
+   */
+  data: string;
 };
 
 /**
@@ -2365,9 +2417,16 @@ type AudienceObject = {
   audienceGroupId: number;
 };
 
+type RedeliveryObject = {
+  type: "redelivery";
+  requestId: string;
+};
+
 export type ReceieptObject =
   | AudienceObject
-  | FilterOperatorObject<AudienceObject>;
+  | RedeliveryObject
+  | FilterOperatorObject<AudienceObject>
+  | FilterOperatorObject<RedeliveryObject>;
 
 type DemographicAge =
   | "age_15"
@@ -2618,6 +2677,15 @@ export type GroupSummaryResponse = {
  */
 export type MembersCountResponse = {
   count: number;
+};
+
+export type GetRichMenuAliasResponse = {
+  richMenuAliasId: string;
+  richMenuId: string;
+};
+
+export type GetRichMenuAliasListResponse = {
+  aliases: GetRichMenuAliasResponse[];
 };
 
 /**